Move empty state from marketplace to shared
