+++ app/src/marketplace/components/Products/NoMyPurchasesView/index.jsx
@@ -3,10 +3,10 @@
 import React from 'react'
 import { Translate, I18n } from 'react-redux-i18n'
 
-import EmptyState from '../../EmptyState'
+import EmptyState from '$shared/components/EmptyState'
 import { main } from '../../../../links'
-import emptyStateIcon from '../../../assets/empty_state_icon.png'
-import emptyStateIcon2x from '../../../assets/empty_state_icon@2x.png'
+import emptyStateIcon from '$shared/assets/images/empty_state_icon.png'
+import emptyStateIcon2x from '$shared/assets/images/empty_state_icon@2x.png'
 
 const NoProductsView = () => (
     <EmptyState
