+++ app/src/shared/assets/images/app_crashed@2x.png
