+++ app/src/shared/assets/images/empty_state_icon.png
