+++ app/src/shared/assets/images/empty_state_icon@2x.png
